% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthro-package.R
\docType{package}
\name{anthro-package}
\alias{anthro}
\alias{anthro-package}
\title{Compute the WHO Child Growth Standards}
\description{
Provides WHO Child Growth Standards (z-scores) with
confidence intervals and standard errors around the
prevalence estimates, taking into account complex sample designs.
More information on the methods is available online:
<\url{https://www.who.int/tools/child-growth-standards}>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WorldHealthOrganization/anthro}
  \item Report bugs at \url{https://github.com/WorldHealthOrganization/anthro/issues}
}

}
\author{
\strong{Maintainer}: Elaine Borghi \email{nfsdata@who.int} [contributor]

Authors:
\itemize{
  \item Dirk Schumacher \email{mail@dirk-schumacher.net}
}

Other contributors:
\itemize{
  \item Jonathan Polonsky \email{polonskyj@who.int} [contributor]
  \item Giovanna Gatica Dominguez \email{gaticag@who.int} [contributor]
  \item World Health Organization [copyright holder]
}

}
