% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthro.R
\docType{package}
\name{anthro}
\alias{anthro}
\alias{package-anthro}
\alias{anthro-package}
\title{A package for the WHO Child Growth Standards}
\description{
A package to compute the WHO Child Growth Standards.
More information on the methods is available online:
<http://www.who.int/childgrowth/standards/en/>.
}
