% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{anthro_api_standardize_oedema_var}
\alias{anthro_api_standardize_oedema_var}
\title{Standardize the Oedema input values}
\usage{
anthro_api_standardize_oedema_var(oedema)
}
\arguments{
\item{oedema}{a vector of values}
}
\description{
Standardize the Oedema input values
}
\note{
This function is meant to be used by other anthro related packages.
It is not advised to use this in you own packages or analysis. If you must
use it, prepare for potential breaking changes in the future.
}
