% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthro.R
\docType{package}
\name{anthro}
\alias{anthro}
\alias{package-anthro}
\title{Compute the WHO Child Growth Standards}
\description{
Provides WHO Child Growth Standards (z-scores) with
confidence intervals and standard errors around the
prevalence estimates, taking into account complex sample designs.
More information on the methods is available online:
<\url{http://www.who.int/childgrowth/standards/en/}>.

Package lifecycle is "maturing" in the tidyverse sense (<\url{https://www.tidyverse.org/lifecycle/}>).
It is aimed to keep the API backward compatible, but changes to the API
might occur.
}
