% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{anthro_api_compute_prevalence}
\alias{anthro_api_compute_prevalence}
\title{Compute prevalence of zscores}
\usage{
anthro_api_compute_prevalence(data, zscores_to_compute, survey_subsets)
}
\arguments{
\item{data}{a data frame containing the underlying data}

\item{zscores_to_compute}{a list of zscore indicators that should be computed.}

\item{survey_subsets}{subsets for which the prevalence values should be computed.}
}
\description{
Compute prevalence of zscores
}
\note{
This function is meant to be used by other anthro related packages.
It is not advised to use this in you own packages or analysis. If you must
use it, prepare for potential breaking changes in the future.
}
