% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{anthro_api_compute_zscore}
\alias{anthro_api_compute_zscore}
\title{Helper function to compute zscores}
\usage{
anthro_api_compute_zscore(y, m, l, s)
}
\arguments{
\item{y}{a numeric vector}

\item{m}{a numeric vector}

\item{l}{a numeric vector}

\item{s}{a numeric vector}
}
\description{
Helper function to compute zscores
}
\note{
This function is meant to be used by other anthro related packages.
It is not advised to use this in you own packages or analysis. If you must
use it, prepare for potential breaking changes in the future.
}
\references{
http://www.who.int/childgrowth/standards/Chap_7.pdf
}
