% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{anthro_api_compute_zscore_adjusted}
\alias{anthro_api_compute_zscore_adjusted}
\title{Helper function to compute the adjusted zscore}
\usage{
anthro_api_compute_zscore_adjusted(y, m, l, s)
}
\arguments{
\item{y}{a numeric vector}

\item{m}{a numeric vector}

\item{l}{a numeric vector}

\item{s}{a numeric vector}
}
\description{
Helper function to compute the adjusted zscore
}
